requirements.txt
Babel==2.5.3
boto3==1.6.0
botocore==1.9.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.11.5
django-activity-stream==0.6.4
django-braces==1.12.0
django-ckeditor==5.3.0
django-cleanup==2.1.0
django-cors-headers==2.1.0
django-countries==5.1.1
django-extensions==2.0.0
django-extra-fields==0.9
django-filter==1.1.0
django-import-export==0.5.1
django-js-asset==1.0.0
django-lazysignup==1.1.2
django-money==0.12.3
django-mptt==0.9.0
django-oauth-toolkit==1.0.0
django-otp==0.4.2
django-otp-twilio==0.4.1
django-phonenumber-field==2.0.0
django-redis==4.8.0
django-redis-cache==1.7.1
django-redis-sessions==0.5.6
django-rest-framework-social-oauth2==1.1.0
django-s3-storage==0.12.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
et-xmlfile==1.0.1
fcm-django==0.2.14
httplib2==0.10.3
idna==2.6
jdcal==1.3
jmespath==0.9.3
oauth2==1.9.0.post1
oauthlib==2.0.6
odfpy==1.3.6
openpyxl==2.5.0
optisorl==0.2.1
phonenumberslite==8.8.1
Pillow==5.0.0
psycopg2==2.7.4
py-moneyed==0.7.0
pyfcm==1.4.3
PyJWT==1.5.3
pymongo==3.5.1
PySocks==1.6.8
python-dateutil==2.6.1
python-google-places==1.4.1
python-magic==0.4.15
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rocketchat-API==0.6.2
s3transfer==0.1.13
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sorl-thumbnail==12.4a1
tablib==0.12.1
twilio==6.10.5
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.22
whitenoise==3.3.0
xlrd==1.1.0
xlwt==1.3.0
django-multiselectfield
boto==2.48.0
django_crontab
django-iban
django-colorful
django-ipware
django-localflavor
dwollav2
django-sslserver



# for windows 
# python-magic==0.4.13
# python-libmagic==0.4.0
# python-magic-bin==0.4.14

